requirements.txt
aiohttp==3.6.2
Pillow==7.0.0
aiogram==2.5.3
peewee==3.13.1
peewee_async==0.6.3a0
aiopg==1.0.0
